+++ app/server/package.json
@@ -41,7 +41,7 @@
     "global": "^4.3.2",
     "regenerator-runtime": "^0.13.3",
     "safe-identifier": "^0.3.1",
-    "ts-dedent": "^1.1.0"
+    "ts-dedent": "^1.1.1"
   },
   "devDependencies": {
     "fs-extra": "^8.0.1"
